source/bilibili_spider/requirements.txt
bilibili-api-python>=14.5.1
requests>=2.25.1
aiohttp>=3.7.4